+++ app.rb
@@ -2,7 +2,6 @@ require_relative 'teacher'
 require_relative 'student'
 require_relative 'book'
 require_relative 'rental'
-require_relative 'manageperson'
 
 class App
   attr_accessor :books, :people, :rentals
@@ -28,6 +27,42 @@ class App
              "Parent Permission: #{person.parent_permission}"
       end
     end
+  end
+
+  def create_teacher
+    print "Teacher's Specialization: "
+    specialization = gets.chomp
+    print "Teacher's age: "
+    age = gets.chomp.to_i
+    print "Teacher's name: "
+    name = gets.chomp
+
+    teacher = Teacher.new(age, name, specialization, parent_permission: true)
+    @people << teacher
+    puts 'You have successfully registered a Teacher'
+  end
+
+  def create_student
+    print 'Age: '
+    age = gets.chomp.to_i
+    print 'Name: '
+    name = gets.chomp
+    print 'Has parent permission? [Y/N]: '
+
+    parent_permission = gets.chomp.downcase
+
+    case parent_permission
+    when 'n'
+      student = Student.new(nil, age, name, parent_permission: false)
+      @people << student
+    when 'y'
+      student = Student.new(nil, age, name, parent_permission: true)
+      @people << student
+    else
+      'You have entered an invalid option'
+    end
+
+    puts 'You have successfully registered a Student'
   end
 
   def create_person
