+++ main.rb
@@ -3,4 +3,5 @@ require_relative 'homepage'
 
 app = App.new
 homepage = HomePage.new(app)
+
 homepage.start
